# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comerc API Documentation",
        "description": "Documentação da API Comerc",
        "contact": {
            "email": "contato@leonardo-marques.tech"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lista todos os produtos",
                "description": "Display a listing of the resource.",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Lista paginada de produtos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Cria um novo produto",
                "description": "Store a newly created resource in storage.",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Mostra detalhes de um produto",
                "description": "Display the specified resource.",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do produto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualiza um produto existente",
                "description": "Update the specified resource in storage.",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Deleta um produto",
                "description": "Remove the specified resource from storage.",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto deletado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "required": [
                    "id",
                    "name",
                    "price",
                    "photo",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Produto A"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99
                    },
                    "photo": {
                        "type": "string",
                        "example": "products_photos/abcd1234.jpg"
                    },
                    "type": {
                        "type": "string",
                        "example": "eletrônico",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-28T20:00:00Z",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-29T15:30:00Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductCreateRequest": {
                "required": [
                    "name",
                    "price",
                    "photo"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Produto A"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 19.99
                    },
                    "photo": {
                        "description": "Imagem do produto (jpeg, png, jpg, gif, svg), max 2048KB",
                        "type": "string",
                        "format": "binary"
                    },
                    "type": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "eletrônico",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Produto A"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 19.99
                    },
                    "photo": {
                        "description": "Imagem do produto (jpeg, png, jpg, gif, svg), max 2048KB",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "eletrônico",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}